Fixed some tests, added option to cancel order
Some tests were not working because they had to away async in the methods used to retrieve data.
Added option for admin to cancel order when the order has been deliver but the customer wishes to cancel it.
If status of the order diff. from pending, customer can still cancel but he will not delete it, it will be up to the admin to decide.
Added phone number and name of the person who placed the order . This has been added to the view